+++ src/Datatypes/DataVendedor.cpp
@@ -2,11 +2,11 @@
 
 DataVendedor::DataVendedor(){};
 
-DataVendedor::DataVendedor(string _nickname, string _contrasena, DTFecha _fecha, int _RUT): DataUsuario(_nickname, _contrasena, _fecha), RUT(_RUT){};
+DataVendedor::DataVendedor(string _nickname, string _contrasena, DTFecha _fecha, long int _RUT): DataUsuario(_nickname, _contrasena, _fecha), RUT(_RUT){};
 
 DataVendedor::~DataVendedor(){};
 
-int DataVendedor::getRUT(){
+long int DataVendedor::getRUT(){
     return RUT;
 }
